move routes to ridewithgps collection
The routes page was a bit neglected, and the strava routes and embedded
gpx files had diverged. In a bid to make the routes page easier to
maintain, move from storing the files locally to a ridewithgps
collection. Currently this collection is tied to Leo's ridewithgps
account, but we could move to a club account if we wanted.

You can add any ridewithgps route to collections, but to ensure that
routes don't disappear I recommend any routes we want to add we make a
copy and use that, and then pledge not to edit it again.