make route generation script take list of URLs rather than user profile
the USER_ID.json endpoint only returns the 20 most recent routes from a
user profile, so just hard code the links to all the routes we want to
add. Tweaked the script slightly, as we no longer have access to the
name up-front. Also sort the output list, to reduce diff size when
re-running the script.

also, added the generated route file to linter ignore as it generated as
json with double quotes and standard freaked out